// src/lib/gameliftstreams-1.1.0.d.ts
export type ApplicationStats = {
	/** Normalized application CPU utilization */
	cpuUtilizationNormalized?: number;
	/** Normalized application memory utilization */
	memoryUtilizationNormalized?: number;
	/** Application memory utilization in mebibytes (MiB) */
	memoryUtilizationMb?: number;
	[key: string]: number | undefined;
};
/**
 * Performance stats for the shared compute that is serving the current session
 * On multi-tenant stream groups, this compute may be shared among multiple sessions
 */
export type SharedSystemStats = {
	/** System CPU utilization as percentage (0-100) */
	cpuUtilization?: number;
	/** System Memory utilization as percentage (0-100) */
	memoryUtilization?: number;
	/** System Memory utilization in mebibytes (MiB) */
	memoryUtilizationMb?: number;
	/** System GPU utilization as percentage (0-100) */
	gpuUtilization?: number;
	/** System VRAM utilization as percentage (0-100) */
	vramUtilization?: number;
	/** System VRAM utilization in mebibytes (MiB) */
	vramUtilizationMb?: number;
	[key: string]: number | undefined;
};
/**
 * Provides both application-level and system-level performance stats
 */
export interface PerformanceStats {
	/** Timestamp representing when the stats were captured */
	timestamp: Date;
	/** Performance stats for the application running in the current session */
	application: ApplicationStats;
	/** Performance stats for the shared compute that is serving the current session */
	system: SharedSystemStats;
}
export interface IClientConnection {
	/**
	 * The connectionState callback is triggered whenever the RTCPeerConnection connectionState value changes.
	 * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection}
	 * @param state - "connecting" | "connected" | "disconnected" | "failed" | "closed"
	 */
	connectionState?: (state: string) => void;
	/**
	 * The channelError callback is triggered whenever there is an internal WebRTC data channel error.
	 * Any such errors are usually unrecoverable and your code should terminate the streaming session.
	 * @param e - error object reported by browser WebRTC implementation
	 */
	channelError?: (e: any) => void;
	/**
	 * The serverDisconnect callback is triggered when the streaming server initiates a disconnection.
	 * @param reasoncode - "idle" | "terminated" | (other strings reserved for future use)
	 */
	serverDisconnect?: (reasoncode: string) => void;
	/**
	 * The applicationMessage callback is triggered whenever a message is received from the application through the data channel.
	 * @param message - message received from application (as byte array)
	 */
	applicationMessage?: (message: Uint8Array) => void;
	/**
	 * The performanceStats callback is triggered whenever the latest performance stats for the current session are received
	 * Stream Session must be configured to share performance stats
	 * @param PerformanceStats - Snapshot of application-level and system-level performance stats
	 */
	performanceStats?: (stats: PerformanceStats) => void;
}
export interface GamepadFilterResult {
	/**
	 * An array of button indices that should be suppressed when forwarding gamepad input
	 * to the stream. Each array element should be in the [0, 16] range according to the
	 * Standard gamepad mapping, as defined in the W3C Gamepad specification.
	 * @see {@link https://w3c.github.io/gamepad/#remapping}
	 */
	suppressButtonIndexes: ReadonlyArray<number>;
}
export interface InputConfiguration {
	/**
	 * Automatically listen for and forward keyboard events when input is attached
	 * @defaultValue true
	 */
	autoKeyboard?: boolean;
	/**
	 * Automatically listen for and forward mouse events
	 * @defaultValue true
	 */
	autoMouse?: boolean;
	/**
	 * Automatically look for connected gamepads and forward gamepad input every frame
	 * @defaultValue true
	 */
	autoGamepad?: boolean;
	/**
	 * Reset all input device state whenever detachInput is called. If false, keyboard and mouse
	 * buttons will be frozen in their current state (down vs up) when detachInput() is called.
	 * @defaultValue true
	 */
	resetOnDetach?: boolean;
	/**
	 * Detaches input devices automatically when the window loses focus, and restores input
	 * devices when the window regains focus. If false, Amazon GameLift Streams ignores window focus and you
	 * must call detachInput() manually to stop tracking user input devices and attachInput()
	 * to resume. This setting combines with resetOnDetach to automatically release any held
	 * keys or buttons when window focus is lost.
	 * @defaultValue true
	 */
	trackWindowFocus?: boolean;
	/**
	 * If set, the keyboardFilter callback function is used to filter every keyboard event.
	 * The callback should return 'false' to prevent the event from being forwarded to the stream.
	 * @defaultValue null
	 */
	keyboardFilter?: null | ((event: KeyboardEvent) => boolean);
	/**
	 * If set, the mouseFilter callback function is used to filter every mouse event.
	 * The callback should return 'false' to prevent the event from being forwarded to the stream.
	 * @defaultValue null
	 */
	mouseFilter?: null | ((event: MouseEvent) => boolean);
	/**
	 * If set, the gamepadFilter callback function is used to filter gamepad inputs.
	 * The callback should return 'false' to completely suppress the gamepad, or a
	 * GamepadFilterResult object to suppress only individual buttons.
	 * @defaultValue null
	 */
	gamepadFilter?: null | ((event: Gamepad) => boolean | GamepadFilterResult);
	/**
	 * Allow haptic feedback events (eg, controller vibration) to be triggered by the stream
	 * @defaultValue true
	 */
	hapticFeedback?: boolean;
	/**
	 * Allow mouse cursor shape to be controlled by the stream
	 * Options:
	 *   true - the local cursor replicates the remote cursor shape on the stream host
	 *   false - the local cursor is never modified or hidden
	 *   'visibility' - the local cursor is never modified, but it can be hidden
	 *   function - the local cursor is never modified, but the function is invoked with a CSS style string representing the remote cursor whenever the remote cursor changes
	 * @defaultValue 'visibility'
	 */
	setCursor?: true | false | "visibility" | ((cursorStyle: string) => void);
	/**
	 * Automatically capture the mouse (hiding the cursor and locking the mouse to the current window)
	 * whenever the remote cursor has been made invisible on the stream host
	 * Options:
	 *   true - the mouse is always captured
	 *   false - the mouse is never captured
	 *   'fullscreen' - the mouse is only captured if the video element is fullscreen
	 *
	 * This is often helpful for "first-person" games which use the mouse for aiming.
	 *
	 * This has no effect in the Safari browser or on the iOS platform due to platform limitations.
	 * @defaultValue 'fullscreen'
	 */
	autoPointerLock?: true | false | "fullscreen";
}
/**
 * Amazon GameLift Streams constructor configuration object
 */
export interface GameLiftStreamsArgs {
	/** HTML video element which will be used to display the stream */
	videoElement: HTMLVideoElement;
	/** HTML audio element which will be used to play the stream audio*/
	audioElement: HTMLAudioElement;
	/** Mouse/keyboard/gamepad configuration properties */
	inputConfiguration?: InputConfiguration;
	/** Connection callbacks */
	clientConnection?: IClientConnection;
	/** Stream configuration properties */
	streamConfiguration?: GameLiftStreamsStreamConfiguration;
}
/**
 * Amazon GameLift Streams stream configuration properties
 */
export interface GameLiftStreamsStreamConfiguration {
	/** Enable audio stream (default: true) */
	enableAudio?: boolean;
	/** Maximum client download bitrate, as kilobits-per-second (default: undefined) */
	maximumKbps?: number;
	/** Maximum client-supported video resolution, regardless of game resolution (default: 1080p) */
	maximumResolution?: "1080p" | "720p" | "540p";
	/**
	 * [EXPERIMENTAL] Client-supplied ICE server information to include in WebRTC connection negotiation. This
	 * setting is experimental and may be removed in the future. It is not recommended outside of specific
	 * enterprise scenarios where clients are known to be behind a restrictive firewall. (default: undefined)
	 */
	iceServers?: GameLiftStreamsIceServerConfiguration[] | (() => GameLiftStreamsIceServerConfiguration[]) | (() => Promise<GameLiftStreamsIceServerConfiguration[]>);
	/**
	 * [EXPERIMENTAL] Force negotiation of a specific WebRTC video codec configuration. This setting is
	 * experimental and may be removed in the future. It should only be set as part of a customer support
	 * investigation into client compatibility with new hardware platforms. (default: undefined)
	 */
	forceVideoCodec?: "h264.constrained.31" | "h264.constrained" | "h264.baseline" | "h264.main.31" | "h264.main" | "h264.high";
}
/** Amazon GameLift Streams STUN/TURN server configuration, allowing either 'urls' or 'url' */
export type GameLiftStreamsIceServerConfiguration = RTCIceServer | {
	/** URL representing this server, starting with 'turn:' or 'turns:' or 'stun:' or 'stuns:' */
	url: string;
	/** Username for TURN server authentication */
	username?: string;
	/** Credential (password) for TURN server authentication */
	credential?: string;
};